First commit
The following is the base of the project that I have compleated over the weekend, It contains a simple flutter application that load a map and track the current users location, it also has the directions API and polylines built-in. The server is just a dumy serverr until I can get the users data mapped out on the app and start sending it over to the server to handle.

Signed-off-by: Hwessels <Hwessels@student.wethinkcode.co.za>